call and jump -> branch
