+++ Makefile
@@ -15,8 +15,7 @@ OBJECTS = \
 	literal.o \
 	mnemonic.o \
 	mnemonic_data.o \
-	mnemonic_call.o \
-	mnemonic_jump.o \
+	mnemonic_branch.o \
 	mnemonic_io.o \
 	mnemonic_load.o \
 	mnemonic_calc.o \
@@ -89,10 +88,7 @@ mnemonic.o: src/mnemonic.cpp ${HEADERS} src/mnemonic.h
 mnemonic_data.o: src/mnemonic_data.cpp ${HEADERS} src/mnemonic.h
 	${CPP} ${CPPFLAGS} -c $< -o $@
 
-mnemonic_jump.o: src/mnemonic_jump.cpp ${HEADERS} src/mnemonic.h
-	${CPP} ${CPPFLAGS} -c $< -o $@
-
-mnemonic_call.o: src/mnemonic_call.cpp ${HEADERS} src/mnemonic.h
+mnemonic_branch.o: src/mnemonic_branch.cpp ${HEADERS} src/mnemonic.h
 	${CPP} ${CPPFLAGS} -c $< -o $@
 
 mnemonic_io.o: src/mnemonic_io.cpp ${HEADERS} src/mnemonic.h
